Refactor out magic numbers in LL2 title.cpp (#11)
The LL2 title uses a lot of magic numbers in the flow control for the menu and character creation dialog. This means if you want to edit the title screen, you have to comb through the code to make sure you set the right numbers in the right places.

This PR `#defines` a bunch of macros to be used in the place of the magic numbers, so that the title screen can be modded in an easier way. This includes making sure the modifier selection still works if you change their subcursor number, and there is enforcement via static assert that the modifier numbers are contiguous so that this change works.