Improve brain and candle auto-counting
Firstly, count monsters which have exactly one baby on death as double,
since they will drop their item exactly twice. For example, a Zomboni
with a brain should count for 3 total brains, not 2. Same for candles.

Second, if the Brain item is changed to give >1 brain, that should be
counted in the math for zombies that auto-drop brains.