Supreme: Add bullet selection menu (#17)
The UX for selecting bullets while editing specials is fairly clunky: 
the user must click through individual bullet types until they find the 
one they want.

This can be improved by adding a menu to select bullets. All bullets 
are presented in a basic list, allowing for faster selection.

The selector is designed to be automatically populated from the bullet 
enum, pulling bullet names from the list in `specialedit.cpp`. A 
`NUM_BULLETS` value has been added to the end of the bullet type list 
to enable this. For the moment, there's no handling for if the number 
of bullet types becomes high enough to create a fifth column of 
buttons, but that would require many more bullet types to be added.