LL: Add character selection to new game screen (#19)
Allows the character to be selected when starting a new normal or remix game, for convenience. This is done with some helper functions and maps that set the appropriate merit badge effects when called. The player can only select characters that are unlocked, and the selection doesn't appear at all if no characters have been unlocked.