Fix LL's Hardcore Mode
In 2.0, the "Quit" option would prompt for a slot to save but instead
*load* that slot, making Hardcore Mode actually unsaveable. This is
fixed by showing the save menu instead.

However, it was also possible to simply alternate your hardcore saves
between two slots, only one of which would be deleted on death. Instead,
show a single "Save & Quit" option (regardless of Save Everywhere),
which only prompts for which slot to save to for the first save of a new
game, otherwise overwriting the loaded save.

HamSandwich also didn't properly delete the save file on death.