SWC: Fix blank lines in text files repeating the previous line
Introduced in 7169378dbaacb40c6e2ef13cdd196db0f1e71b30.

VariableMsg has long had a bug where passing the empty string would
return whatever the previous buffer was, triggerable for example by
showing a blank message or having a map with a blank name.

Using fgets, the line buffer included the '\n' and so was not empty, but
getline excludes the terminator so the buffer was empty, triggering this
behavior of VariableMsg where it wasn't before.

Let's just fix VariableMsg once and for all.