Compile our own SDL on Linux
This lets us configure which dependencies are linked how, so we don't
end up linking to libFLAC, libwebp, and so on bloating our Itch builds
just because that's how newer versions of the Steam Runtime SDK
configure things.

Also switches to STB instead of reference parsers for png, jpg, and ogg.